requirements.txt
# Web Scraping
scrapy==2.8.0
beautifulsoup4>=4.12.0
selenium==4.10.0
requests>=2.31.0

# Database
psycopg2-binary
redis==4.5.5

# Processing
spacy>=3.6.0
transformers>=4.30.0
scikit-learn>=1.3.0
celery==5.3.1

# Frontend
streamlit>=1.28.0
elasticsearch==8.9.0

# Infrastructure
docker==6.1.3
kubernetes==26.1.0
prometheus-client==0.17.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.0
feedparser>=6.0.0
python-dateutil>=2.8.0
lxml>=4.9.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
nltk>=3.8.0 